chore: add back default props to Preact
